requirements.txt
Flask==2.3.3
python-nmap
scapy